// Blockchain-Carpool/typechain-types/CarPooling.ts
export declare namespace CarPooling {
  export type DriverStruct = { isRegistered: boolean; hasRide: boolean };

  export type DriverStructOutput = [isRegistered: boolean, hasRide: boolean] & {
    isRegistered: boolean;
    hasRide: boolean;
  };

  export type PassengerStruct = { isRegistered: boolean; hasRide: boolean };

  export type PassengerStructOutput = [
    isRegistered: boolean,
    hasRide: boolean
  ] & { isRegistered: boolean; hasRide: boolean };

  export type RideStruct = {
    rideId: BigNumberish;
    driver: AddressLike;
    travelTime: BigNumberish;
    availableSeats: BigNumberish;
    totalSeats: BigNumberish;
    seatPrice: BigNumberish;
    origin: BigNumberish;
    destination: BigNumberish;
    status: BigNumberish;
    passengerAddr: AddressLike[];
  };

  export type RideStructOutput = [
    rideId: bigint,
    driver: string,
    travelTime: bigint,
    availableSeats: bigint,
    totalSeats: bigint,
    seatPrice: bigint,
    origin: bigint,
    destination: bigint,
    status: bigint,
    passengerAddr: string[]
  ] & {
    rideId: bigint;
    driver: string;
    travelTime: bigint;
    availableSeats: bigint;
    totalSeats: bigint;
    seatPrice: bigint;
    origin: bigint;
    destination: bigint;
    status: bigint;
    passengerAddr: string[];
  };
}